requirements.txt
flask
python-dotenv
openai>=1.54.0
gunicorn
